requirements_prod.txt
numpy==1.23.5
pandas==2.2.2
requests==2.32.3
scikit-learn==1.5.1
fastapi==0.112.1
uvicorn==0.30.6
google-cloud-storage==2.14.0
gcloud
h5py==3.10.0
python-dotenv